client: Shuffle some functions around.
Move some useful functions from cmd and webserver to a new helper file
in core and dex/encode so that they can be used in other places. The
moved functions are used to parse file paths and orders and are helpful
for creating comma separated values.
