+++ client/core/helpers_test.go
@@ -1,19 +1,17 @@
-// This code is available on the terms of the project LICENSE.md file,
-// also available online at https://blueoakcouncil.org/license/1.0.0.
+//go:build !harness
+// +build !harness
 
-package webserver
+package core
 
 import (
 	"testing"
 	"time"
-
-	"decred.org/dcrdex/client/core"
 )
 
 func Test_matchReader_TimeString(t *testing.T) {
 	stamp := uint64(1607189329)
 	mr := &matchReader{
-		Match: &core.Match{
+		Match: &Match{
 			Stamp: stamp * 1000,
 		},
 	}
