+++ dex/testing/loadbot/go.mod
@@ -7,5 +7,4 @@ replace decred.org/dcrdex => ../../../
 require (
 	decred.org/dcrdex v0.1.3
 	github.com/Shopify/toxiproxy v2.1.4+incompatible
-	golang.org/x/text v0.3.6
 )
